fix(*): organize files
