+++ packages/entities/entities-plugins/src/components/select/PluginCardSkeleton.vue
